pyproject.toml
[build-system]
requires = ["maturin>=1.5,<2.0"]
build-backend = "maturin"

[project]
name = "quinoa"
version = "0.1.0"
description = "A meeting recording and transcription app for Linux"
requires-python = ">=3.12"
dependencies = [
    "PyQt6>=6.6.0",
    "google-genai>=0.3.0",
    "keyring>=24.0.0",
    "pydantic>=2.0.0",
    "markdown>=3.5.0",
    "markdownify>=0.11.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pytest>=8.0.0",
]

[tool.maturin]
features = ["pyo3/extension-module", "real-audio"]
module-name = "quinoa_audio"
manifest-path = "quinoa_audio/Cargo.toml"

# Ruff - fast Python linter and formatter
[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort (import sorting)
    "B",      # flake8-bugbear (common bugs, bare except)
    "T20",    # flake8-print (no print statements)
    "UP",     # pyupgrade (modern Python syntax)
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (let formatter handle it)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["T20"]  # Allow print in tests

[tool.ruff.format]
quote-style = "double"

# Mypy - static type checking
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
# Start lenient, tighten over time
ignore_missing_imports = true
check_untyped_defs = false
